Update Maven dependencies. Also update plugin dependencies.
Add a prerequisite for Maven version.
The output of the Maven versions plugin is more readable
with a minimal Maven version and 3.1 looks like a reasonable base
version.

Exclude the Guava classes from the checkstyle scan as they don't
respect the new checkstyle rules but we want to keep them in sync
with Guava's original classes.