Fix parsing standard-conforming strings
It's impossible to unambiguously parse a string when two different ways
of escaping are possible: standard (where a single quote is escaped with
another single quote) and non-standard (where a single quote is escaped
with a backslash). One example of an ambiguous string is `\''`. In order
to parse it correctly parsing code needs to know whether current string
literal is standard-conforming or not.

Fixes #254