Make BasicContext.loadLocale() handle Windows locales.
The locales that PostgreSQL uses on Windows are different from their
*nix counterparts. They look like 'English_United States.1252' whereas
a *nix locale looks like 'en_US.UTF-8'.

Attempting to parse a Windows locale with the current code results in
an error:

Caused by: java.util.IllformedLocaleException: Ill-formed region: United States [at index 0]
        at java.util.Locale$Builder.setRegion(Locale.java:2495)
        at com.impossibl.postgres.system.BasicContext.loadLocale(BasicContext.java:285)
        at com.impossibl.postgres.system.BasicContext.init(BasicContext.java:265)
        at com.impossibl.postgres.jdbc.PGConnectionImpl.init(PGConnectionImpl.java:249)
        at com.impossibl.postgres.jdbc.ConnectionUtil.createConnection(ConnectionUtil.java:182)
        at com.impossibl.postgres.jdbc.AbstractDataSource.createConnection(AbstractDataSource.java:747)
        at com.impossibl.postgres.jdbc.PGDataSource.getConnection(PGDataSource.java:66)
        ... 6 more

This is because Java is not able to parse a Windows locale. So, we
create a mapping from Windows to *nix locales in utils/Locales.java
and use the utility function getJavaCompatibleLocale(String winLocale)
to see if the locale returned by PostgreSQL matches any known Windows
locale.

The list of locales in utils/Locales.java is not exhaustive, but it
should be sufficient at least as a start.