Server connection is now an asynchronous engine
The “protocol” (aka connection) has been replaced with a completely asynchronous system.

Now the client submits requests, providing a handler, and the server fulfills them in fifo order. This helps with the previous performance issues an ensures the code can support an ADBA interface when it’s ready to be implemented.

# Conflicts:
#	src/main/java/com/impossibl/postgres/jdbc/PGBuffersStruct.java