Add statement description cache.
Adds an un-typed statement description cache to the connectoin. Describing non-cached statements repeatedly for parameter type information is costly. Preparing everything in advance is costly and resource consuming. This cache stores the results of a server parse but does not prepare the query on the server.

Given it’s a very small resource load the cache is quite large (although not unbounded) and the prepared statement can now use the cached descriptions from in the prepared statement cache as well.