Rework 2 -  Async Boogaloo (#350)
* Server connection is now an asynchronous engine

The “protocol” (aka connection) has been replaced with a completely asynchronous system.

Now the client submits requests, providing a handler, and the server fulfills them in fifo order. This helps with the previous performance issues an ensures the code can support an ADBA interface when it’s ready to be implemented.

* Support parameters in DirectQuery

It uses the extended query protocol but parses every time.

* Fix batch commit statement name

* Optimize result field format encoding

* Skip writing formats for lazy requests

They should have no results to worry about.

* BasicContext.PreparedQuery to QueryDescription

It duplicates the name of a new JDBC class that it fits better with.

* Enhance generic server connection interface

Allows properly hiding internal components that don’t need to be accessible otherwise.

* Heat based statement caching

* Use text format for string types

* New engine passes the entire test suite

* Add statement description cache.

Adds an un-typed statement description cache to the connectoin. Describing non-cached statements repeatedly for parameter type information is costly. Preparing everything in advance is costly and resource consuming. This cache stores the results of a server parse but does not prepare the query on the server.

Given it’s a very small resource load the cache is quite large (although not unbounded) and the prepared statement can now use the cached descriptions from in the prepared statement cache as well.

* Cleanup ProtocolChannel

* Remove logging (message tracing is much more concise and easier to read)
* Always write messages using netty’s `voidPromise()` since we never use the results

* Refine & document RequestExecutor interface

Separated out simple and extended query type handlers to make it clear when a `handleReady` or `handleSuspend` callback can be received.

* Refactor all dispatching logic/state into handler

* Make lazy execute perform a sync

This ensures transaction status is updated prior to the next request being sent.

* Cotext shutdown waits for connection shutdown

This ensures state is reported correctly outside

* Fix names in startup handler

* Update shade plugin

* Ensure `abort` uses a valid socket address

* Respect network timeout during close

* Make connection test more reliable

* Fix shutdown not releasing shared connection state

* ServerConnection shutdown waits for `kill`

`ServerConnect.shutdown` now waits for writing the termination packet _and_ the kill (which included channel close)

* Remove uncessary scroller cancel & refresh

* Wait for shutdown in `abort`

* Add setting for max I/O threads

* Refine RequestExecutor interface

The request executor interface itself is now more specific about what callbacks can and will happen for any given request.  Also, the documentation now better describes what requests should be followed by other requests (e.g. suspended requests *should* use `finish` not `close`.

* Allow multiple calls of ServerConnection.shutdown

This ensures each call gets a valid future they can wait on an be notified when the shutdown has actually completed.

* Remove unused QueryTimeoutException

* Remove unecessary functional interface

* Cleanup unused method

* Add network timeout to connection abort test

* Reduce TypeRef garbage

Also, fixes (by removing) “Refs” from parameter names

* Remove little used `ResultBatch.Row`

* Cache referenced types in result batches

* Fix excess flushes from dispatch handler

* Add sql query tracing

Tracing is done in the request executor. For simple queries it shows the text. For prepares and prepared queries it shows the text and statement name.

* Fix checkstyle errors