Default to using native if the libraries are avail
If the native netty libraries are on the classpath for the current OS we default to using them, and falling back to NIO; if somebody put them there they probably want to use them.

This lessens the likelihood that somebody will _forget_ to enable and mix them with domain sockets which would create 2 separate thread pools.