Revamp how settings are defined, found & stored
Added a `Setting` object which defines information related to each setting (description, name, alternate names, default value, etc).  Settings can even define a dynamic source for it’s default value. This will ensure that we always know everything about each setting we define and where/how the default values are determined.  Removes a lot of parameter/setting copying code to make sure things are named right and in the right properties instance.

Allowing alternate names for a setting allows changing them in the future. This isn’t just a “nice to have” currently we use incorrect names for datasource properties; this allows a backwards compatible fix.

Also, added `Settings` which is much like `java.util.Properties` in that it stores everything in text but it allows access via `Setting` definitions.  `Settings` logs warnings for unrecognized names and logs errors for invalid values that are stored.