Make `strict-mode` strict again
Until now enabling `strict-mode` made the driver _less_ compliant. It’s fixed and `strict-mode` is now on by default.  Also, I’ve documented in the setting what behavior changes and when.

#### getColumnName/getColumnLabel
The only place strict mode was still used was in the discrepency between `ResultSetMetaData`’s `getColumnName` and `getColumnLabel`.  According to the spec `getColumnName` should return the table name (other drivers seem to agree at this point).  We had it reversed, strict mode made it return a column label first (if available) then fell back to the table name. Running in `strict-mode` was  “more compatible” / “less strict”.

I’ve now corrected this case and added full tests for both cases. I’v also enabled `strict-mode` by default. This shouldn’t be an issue because it was disabled before the logic was inverted.

#### BatchUpdateException
Along with fixing the previous case I added back another area I knowing defied the standard with regard to `BatchUpdateException` including and EXECUTION_FAILED flag.  In this case we were just “non-compliant” and included the flag. Since the spec is clear about this I re-instated the compliant behavior and stopped adding the flags when in `strict-mode` and add the flag when `strict-mode` is disabled. This change _does_ represent a reversal as, per default, it is now compliant.