Switch to Gradle build (#369)
* Switch to Gradle build

The switch was made because maven’s support for building a generator library and/or annotation processor in the same project is lacking and causes update problems. Gradle has `buildSrc` and is designed to work with this exact scenario.

### Docker
Docker support has been added to the build scripts. It allows testing against a “default” version of PostgreSQL (currently 11) with normal gradle command.

#### Testing specific or all server versions
Testing against a specific server version can be done using the `testPostgres<version>` family of tasks. Finally, testing against all versions of a server is supported using the `testAllPostgresVersions` task.

### OSGI
The entire build is converted with the exception of `OSGI` packaging. The maven build’s OSGI packaging was already incorrect and produced invalid bundles. This will be added back in an upcoming change that tackles it correctly.

* Fix deprecation use warnings in tests