Fix server connection listener being collected early (#381)
There was no hard reference to the `ServerConnection.Listener` created in `BasicContext` whic caused it to be collected early.

This adds a hard reference inside `BasicContext` to keep it alive. This caused a failure in LeakTest due to a mix of self references from anonymous classes and weak references; though the exact cause was mysterious.

The anonymous classes were for caching and all moved into a new `CacheMap` class and weak references were used when back connections were required to fix the LeakTest and simplify code.