Pipeline `executeBatch` requests (#384)
* Add cleaning to githubRelease task

* Pipeline `executeBatch` requests

This dramatically reduces the overhead in batch executions without compromising functionality. My preliminary tests show about 70% reduction in batch execution.

The implementation continuously checks the status of outstanding requests, completing them and checking for failures. This ensures that as soon as the driver knows about an error, it stops sending requests to the server.