Interval overhaul
## Interval class
The `Interval` type has been completely replaced with a verions based on Java 8’s `java.time` package.  The file is based on the fully featured `PeriodDuration` class from [three-ten](https://github.com/ThreeTen/threeten-extra).

## Server/Client parsing
Like Numeric & Date parsing/formatting the “server” parsing is matched to the server based on the `IntervalStyle` server setting.  The client parsing is independently set to ISO.  This means that all JDBC text input for intervals is in ISO 8601 format and doesn’t accept Postgres’s interval format.

To pass data to the server in Postgres format it can be done be passing it as text and casting it to an interval (or relying on implicit casting) on the server.