Simplify SASL/SCRAM interface & usage
All of the logic & decision making has been moved into these implementation classes.  A `ScramSessionFactory` is still built using a builder but the consumer now supplies readily available information (e.g. the server’s advertised mechanisms and channel-bind method) and relies on the builder and client to make the correct choices.

`ScramSessionFactory` is the new name for `ScramClient`.

A `ScramSession` is now a fully encapsulated authentication driver. All of the data and logic required to perform authentication is contained within. It provides a simple interface to drive the authentication process with very little decision making and almost no knowledge of the protocol.